@@ -16,12 +16,14 @@
     "node": ">=6.0.0"
   },
   "dependencies": {
-    "crc": "3.4.0",
-    "debug": "^3.1.0"
+    "crc": "4.3.2",
+    "debug": "^4.4.0",
+    "buffer": "^6.0.3",
+    "eventemitter3": "^5.0.1"
   },
   "devDependencies": {
     "@types/crc": "^3.4.0",
-    "@types/debug": "^4.1.4",
+    "@types/debug": "^4.1.12",
     "@types/mocha": "^5.2.7",
     "@types/node": "^12.0.10",
     "commander": "^3.0.2",
@@ -29,9 +31,8 @@
     "nyc": "^13.1.0",
     "serialport": "^13.0.0",
     "sinon": "2.2.0",
-    "tslint": "^5.20.0",
-    "tslint-config-standard": "^8.0.1",
-    "typescript": "^3.5.2"
+    "tslint": "^6.1.3",
+    "typescript": "^5.7.3"
   },
   "scripts": {
     "test": "npm run lint && npm run build && npx mocha test/*.test.js",
