@@ -1,8 +1,9 @@
-import Debug = require('debug'); const debug = Debug('rtu-request')
-import CRC = require('crc')
-import ModbusAbstractRequest from './abstract-request.js'
-import ModbusRequestBody from './request/request-body.js'
-import RequestFactory from './request/request-factory.js'
+import Debug from 'debug'; const debug = Debug('rtu-request')
+import CRC from 'crc'
+import ModbusAbstractRequest from './abstract-request'
+import ModbusRequestBody from './request/request-body'
+import RequestFactory from './request/request-factory'
+import { Buffer } from 'buffer';
 
 /**
  *
