@@ -1,12 +1,12 @@
 
-import ModbusAbstractRequest from './abstract-request'
-import { RequestToResponse } from './request-response-map'
-import ModbusRTURequest from './rtu-request'
-import ModbusTCPRequest from './tcp-request'
+import type ModbusAbstractRequest from './abstract-request'
+import type { RequestToResponse } from './request-response-map'
+import type ModbusRTURequest from './rtu-request'
+import type ModbusTCPRequest from './tcp-request'
 import { UserRequestError } from './user-request-error'
 import { UserRequestMetrics } from './user-request-metrics'
 
-import Debug = require('debug'); const debug = Debug('user-request')
+import Debug from 'debug'; const debug = Debug('user-request')
 
 export type ModbusRequest = ModbusTCPRequest | ModbusRTURequest
 export interface IUserRequestResolve<Req extends ModbusAbstractRequest> {
