@@ -1,11 +1,11 @@
 
-import MBClient from './modbus-client.js'
-import ModbusRTUClientRequestHandler from './rtu-client-request-handler.js'
-import ModbusRTUClientResponseHandler from './rtu-client-response-handler.js'
+import MBClient from './modbus-client'
+import ModbusRTUClientRequestHandler from './rtu-client-request-handler'
+import ModbusRTUClientResponseHandler from './rtu-client-response-handler'
 
 import {SerialPort} from 'serialport'
-import ModbusRTURequest from './rtu-request.js'
-import ModbusRTUResponse from './rtu-response.js'
+import ModbusRTURequest from './rtu-request'
+// import ModbusRTUResponse from './rtu-response'
 
 /** This Client musst be initiated with a socket object that implements the event emitter
  * interface and fires a 'data' event with a buffer as a parameter. It also needs to
