@@ -1,5 +1,6 @@
+import { Buffer } from 'buffer'
 import { FC } from '../codes'
-import ModbusRequestBody from './request-body.js'
+import ModbusRequestBody from './request-body'
 
 /** Read Holding Registers Request Body
  * @extends ModbusRequestBody
